// src/app/resources/user-web/footer/interface/footer.interface.ts
export interface FooterDataType {
  logo: string;
  description: string;
  contact: string;
  email: string;
  phone: string;
  location: string;
  copyright: string;
  facebookLink: string;
  telegramLink: string;
  youtubeLink: string;
}

// const footerDataKh: FooterDataType = {
//   logo: '/images/CamCyber_Logo_white.png',
//   description:
//     'ផលិតផលនិងសេវាទំនើបឌីជីថលកម្ពុជា សម្រាប់កុំព្យូទ័រ ឧបករណ៍ និងអនាគត',
//   contact: 'ទំនាក់ទំនង',
//   email: 'info@camcyberdigital.com',
//   phone: '+855 87 955 888',
//   location: 'Phnom Penh',
//   copyright: '2024 CamCyber Digital Tech Team',
//   facebookLink: 'https://www.facebook.com/?_rdc=1&_rdr',
//   telegramLink: '#',
//   youtubeLink: '#',
// };
// const footerDataEn: FooterDataType = {
//   logo: '/images/CamCyber_Logo_white.png',
//   description:
//     'Modern digital products and services in Cambodia for computers, devices, and the future.',
//   contact: 'Contact',
//   email: 'info@camcyberdigital.com',
//   phone: '+855 87 955 888',
//   location: 'Phnom Penh',
//   copyright: '2024 CamCyber Digital Tech Team',
//   facebookLink: 'https://www.facebook.com/?_rdc=1&_rdr',
//   telegramLink: '#',
//   youtubeLink: '#',
// };